backend/requirements.txt
# ============================================
# Smart Competitor Finder - Backend Dependencies
# v2.0 (18/02/2026) — Rimosso sentence-transformers/torch/scikit-learn
# Classificazione competitor ora via OpenAI gpt-4o-mini
# ============================================

# --- Core Framework ---
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings==2.1.0
pydantic==2.5.0
python-multipart

# --- Web Scraping & Automation ---
playwright
playwright-stealth
beautifulsoup4
lxml
requests
aiofiles
httpx
brotli>=1.1.0
brotlicffi>=1.1.0
aiohttp>=3.9.0
fake-useragent
trafilatura

# --- Data Processing ---
numpy==1.26.4
pandas==2.1.4
openpyxl
xlrd
nltk

# --- AI/ML APIs ---
openai

# --- Database (Future Caching) ---
sqlalchemy
alembic

# --- Utilities & Logging ---
structlog

# --- Development & Testing ---
pytest
pytest-asyncio
